+++ packages/dartcv/lib/src/core/keypoint.dart
@@ -14,9 +14,9 @@ import 'base.dart';
 import 'vec.dart';
 
 class KeyPoint extends CvStruct<cvg.KeyPoint> {
-  KeyPoint._(ffi.Pointer<cvg.KeyPoint> ptr, [bool attach = true]) : super.fromPointer(ptr) {
+  KeyPoint._(ffi.Pointer<cvg.KeyPoint> ptr, {bool attach = true}) : super.fromPointer(ptr) {
     if (attach) {
-      finalizer.attach(this, ptr.cast(), detach: this);
+      finalizer.attach(this, ptr.cast(), detach: this, externalSize: ffi.sizeOf<cvg.KeyPoint>());
     }
   }
   factory KeyPoint(double x, double y, double size, double angle, double response, int octave, int classID) {
@@ -33,7 +33,8 @@ class KeyPoint extends CvStruct<cvg.KeyPoint> {
   }
   factory KeyPoint.fromNative(cvg.KeyPoint r) =>
       KeyPoint(r.x, r.y, r.size, r.angle, r.response, r.octave, r.classID);
-  factory KeyPoint.fromPointer(ffi.Pointer<cvg.KeyPoint> p, [bool attach = true]) => KeyPoint._(p, attach);
+  factory KeyPoint.fromPointer(ffi.Pointer<cvg.KeyPoint> p, {bool attach = true}) =>
+      KeyPoint._(p, attach: attach);
 
   static final finalizer = ffi.NativeFinalizer(calloc.nativeFree);
 
@@ -80,13 +81,19 @@ class KeyPoint extends CvStruct<cvg.KeyPoint> {
 }
 
 class VecKeyPoint extends Vec<cvg.VecKeyPoint, KeyPoint> {
-  VecKeyPoint.fromPointer(super.ptr, [bool attach = true]) : super.fromPointer() {
+  VecKeyPoint.fromPointer(super.ptr, {bool attach = true, int? length}) : super.fromPointer() {
     if (attach) {
-      finalizer.attach(this, ptr.cast<ffi.Void>(), detach: this);
+      finalizer.attach(
+        this,
+        ptr.cast<ffi.Void>(),
+        detach: this,
+        externalSize: length == null ? null : length * ffi.sizeOf<cvg.KeyPoint>(),
+      );
     }
   }
 
-  factory VecKeyPoint([int length = 0]) => VecKeyPoint.fromPointer(ccore.std_VecKeyPoint_new(length));
+  factory VecKeyPoint([int length = 0]) =>
+      VecKeyPoint.fromPointer(ccore.std_VecKeyPoint_new(length), length: length);
 
   factory VecKeyPoint.fromList(List<KeyPoint> pts) =>
       VecKeyPoint.generate(pts.length, (i) => pts[i], dispose: false);
@@ -98,7 +105,7 @@ class VecKeyPoint extends Vec<cvg.VecKeyPoint, KeyPoint> {
       ccore.std_VecKeyPoint_set(p, i, v.ref);
       if (dispose) v.dispose();
     }
-    return VecKeyPoint.fromPointer(p);
+    return VecKeyPoint.fromPointer(p, length: length);
   }
 
   static final finalizer = OcvFinalizer<cvg.VecKeyPointPtr>(ccore.addresses.std_VecKeyPoint_free);